pyproject.toml
[project]
name = "kyc-document-processing"
version = "0.1.0"
description = "Traitement automatique de documents KYC avec LLM multimodaux"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Vinorth VARATHARASAN"}
]
keywords = ["kyc", "llm", "document-processing", "multimodal", "gemini", "ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "google-cloud-aiplatform>=1.38.0",
    "google-cloud-documentai>=2.20.0",
    "langchain>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
